Fix HierachyInfo equality and GetHashCode.
When a file had multiple fields annotated with AvaStyledProperty, the generator would try to create multiple files with the same name instead of generating all properties in the same file.

This happened because the GroupBy method failed to group these fields together when using HierarchyInfo as the Key, due to the ImmutableArray in it. Overriding the GetHashCode and Equals methods fixes this.